+++ common/src/main/java/net/wizards/client/WizardsClientMod.java
@@ -1,8 +1,8 @@
 package net.wizards.client;
 
-import mod.azure.azurelibarmor.common.render.armor.AzArmorRenderer;
-import mod.azure.azurelibarmor.common.render.armor.AzArmorRendererRegistry;
 import net.minecraft.util.Identifier;
+import net.rpg_foundation.armor_api.client.ArmorRenderers;
+import net.rpg_foundation.armor_api.client.GeoArmorRenderer;
 import net.spell_engine.api.effect.CustomModelStatusEffect;
 import net.spell_engine.api.effect.CustomParticleStatusEffect;
 import net.spell_engine.api.render.LightEmission;
@@ -18,7 +18,6 @@ import net.wizards.effect.WizardsEffects;
 import net.wizards.item.WizardArmors;
 
 import java.util.List;
-import java.util.function.Supplier;
 
 public class WizardsClientMod {
     public static void init() {
@@ -27,13 +26,13 @@ public class WizardsClientMod {
         CustomParticleStatusEffect.register(WizardsEffects.frozen.effect, new FrozenParticles(2));
         CustomModelStatusEffect.register(WizardsEffects.frozen.effect, frozenModelFxRenderer());
         CustomModelStatusEffect.register(WizardsEffects.frostShield.effect, new FrostShieldRenderer());
-        registerArmorRenderer(WizardArmors.wizardRobeSet, WizardArmorRenderer::wizard);
-        registerArmorRenderer(WizardArmors.arcaneRobeSet, WizardArmorRenderer::arcane);
-        registerArmorRenderer(WizardArmors.fireRobeSet, WizardArmorRenderer::fire);
-        registerArmorRenderer(WizardArmors.frostRobeSet, WizardArmorRenderer::frost);
-        registerArmorRenderer(WizardArmors.netherite_arcane, WizardArmorRenderer::netheriteArcane);
-        registerArmorRenderer(WizardArmors.netherite_fire, WizardArmorRenderer::netheriteFire);
-        registerArmorRenderer(WizardArmors.netherite_frost, WizardArmorRenderer::netheriteFrost);
+        registerArmorRenderer(WizardArmors.wizardRobeSet, WizardArmorRenderer.wizard());
+        registerArmorRenderer(WizardArmors.arcaneRobeSet, WizardArmorRenderer.arcane());
+        registerArmorRenderer(WizardArmors.fireRobeSet, WizardArmorRenderer.fire());
+        registerArmorRenderer(WizardArmors.frostRobeSet, WizardArmorRenderer.frost());
+        registerArmorRenderer(WizardArmors.netherite_arcane, WizardArmorRenderer.netheriteArcane());
+        registerArmorRenderer(WizardArmors.netherite_fire, WizardArmorRenderer.netheriteFire());
+        registerArmorRenderer(WizardArmors.netherite_frost, WizardArmorRenderer.netheriteFrost());
 
         CustomParticleStatusEffect.register(WizardsEffects.evocation.effect, new EvocationParticles());
 
@@ -68,7 +67,7 @@ public class WizardsClientMod {
                 .entityScaling(ModelFxEffectRenderer.SizeAxis.WIDTH, 0.5F);
     }
 
-    private static void registerArmorRenderer(Armor.Set set, Supplier<AzArmorRenderer> armorRendererSupplier) {
-        AzArmorRendererRegistry.register(armorRendererSupplier, set.head, set.chest, set.legs, set.feet);
+    private static void registerArmorRenderer(Armor.Set set, GeoArmorRenderer renderer) {
+        ArmorRenderers.register(renderer, set.head, set.chest, set.legs, set.feet);
     }
 }